[compiler] Allow any target-dependent event type
This loosens the mux compiler specification to permit that the event
type used by the DMA builtins is any type chosen by the target, as long
as it's consistent across the module.

The spec was hereto referencing a nebulous `__mux_dma_event_t` type
which was under specified, and our compiler implementation was assuming
an opaque pointer. Notably, our compiler implementation doesn't actually
care what the event type, as it's just a pass-through operand.

Since the event type can be any type, this changes the
ReplaceAsyncCopiesPass to pass the (opaque) event type from the
OpenCL-level builtins to the mux builtins. Thus if the target wants to
use this pass but define their own event type, they can replace that
type module-wide before calling the pass. This is what the default
implementation on LLVM 17 will do, by replacing the "spirv.Event" target
extension type with a concrete type, thus resulting in mux builtins with
that concrete event type.

If the target isn't ready to replace event types, they can still make
use of the replacement pass and have the opaque pointers or target
extension types live on until later - this doesn't break the
specification.

This also takes the opportunity to update the ReplaceAsyncCopiesPass to
account for the fact we're using opaque pointers everywhere now, so we
don't have to do quite so many bitcasts.