[compiler] Enable mux subgroup/workgroup builtins
This enables the use of the mux subgroup and workgroup 'collective'
builtins in the pipeline, replacing the OpenCL ones early on.

The passes which used to care about the OpenCL builtins have been
updated to care about the mux builtins. There is no real change in
behaviour, except in a few cases where OpenCL any/all worked on i32
types but the mux versions work on i1 types: vecz was already
translating between these two types, so there's less to do now.

This should simplify a lot of the code.

* The mux sub-group builtins are defined by DefineMuxBuiltinsPass; there
  is no 'trivial' implementation in the OpenCL headers.
* As such, the LinkBuiltinsPass no longer needs to know whether it's
  'early' or not, as the sub-group builtins are declarations and are
  thus safe to link at any point.
* We can get away with doing a lot less function name demangling,
  mangling and manipulation, as we have exposed enums for the builtin
  IDs.
  * The DegenerateSubGroupPass still does name shenanigans, but we can
    fix that up separately.