[compiler] Start to combine 'ReplaceXXX' passes
We have many passes which replace either generic language builtins
(`ReplaceGroupFuncsPass` & `ReplaceBarriersPass`) or hard-coded OpenCL
builtins (`ReplaceAsyncCopiesPass`, `ReplaceAtomicFuncsPass`,
`ReplaceC11AtomicFunsPass`) with some kind of mux/llvm equivalents.

This is not a scalable approach and is hard to communicate to
users, and so this commit starts to merge these passes into one more
versatile pass. Its job is to ask the language BuiltinInfo which
builtins are to be replaced with mux equivalents, so that the following
compiler passes can correctly understand the program semantics.

This marks the beginning of us moving more things more consistently into
the 'mux' sphere and dropping the support for language builtins at all
stages of compilation.