[compiler] Remove MaterializeAbsentWorkItemBuiltinsPass
This pass was originally necessary because SPIR-V would generate OpenCL
version 2.0+ builtins even for an OpenCL 1.2 mux target. Since these
builtins relied on definitions in the builtins module to map them to mux
builtins, and we didn't link the OpenCL 3.0 builtins module for those
old drivers, they would be 'missing'

However, this pass is no longer necessary. This is because now the compiler
itself maps OpenCL builtins to mux builtins (meaning we don't rely on
linking with the builtins module) and because SPIR-V unconditionally
sets the OpenCL 3.0 metadata, these builtins are picked up even in
OpenCL 1.2 mode.