[compiler] Fold ReplaceAsyncCopiesPass into LowerToMuxBuiltinsPass
The ReplaceAsyncCopiesPass was hard-coded to expect OpenCL builtins,
which is not generic enough and won't work on other types of input IR if
we ever decide to support that.

This folds its job into the newly created LowerToMuxBuiltinsPass, as
it's really the same process. The only difference is that before it was
*defining* these OpenCL builtins with bodies that call out to mux
builtins, whereas now it replaces calls to those builtins with calls to
mux builtins (and some IR to transform the arguments).