[vecz] Vectorize sub-group local ID on top of mux sub-groups
Mux sub-groups represent hardware sub-groups (groups of invocations of
the kernel working together), whereas vecz vectorizes *on top* of those
(groups of work-items working together at a software level on *one*
invocation). This should essentially multiply the apparent sub-group
size by the vectorization factor.

The vectorizer is not currently obeying this paradigm, instead replacing
the mux sub-groups with its own and confusing the concept of a mux
sub-group with a vectorized operation. Since we only support a mux
sub-group size of 1 in all our targets, we can just about get away with
this.

We aim to bring the separate concepts of mux subgroups and vectorized
sub-groups together over a series of commits, and this is the first.

The sub-group local ID is now vectorized over mux sub-groups, by
multiplying the mux sub-group local ID by the vectorization factor, and
adding the step vector to that.