[vecz] Do not vectorize llvm.debug.value intrinsics
We were trying to update dbg.value intrinsics with the vectorized value
and by updating the DIType to a constructed vector type.

Unfortunately we weren't always getting this right, and sometimes
generating invalid debug expressions (noticed on LLVM 17).

I have evaluated the state of our debug info (with LLVM and lldb 16) and
found while I was able to step through kernels with a source view, the
experience is lacking overall and I was unable to inspect any variables
at all, regardless of whether we vectorized or not.

Given the fact that LLVM 17 errors about some invalid debug locations
and the fact we have no clear pathway towards fixing this properly, the
best thing to do is to remove our support for vectorized dbg.value
expressions for now.

We can revisit this with a broader view to improving out debugging
support later on.