[vecz] Vectorize sub-group scans on top of mux sub-groups
The process of vectorizing a sub-group scan is a little more complicated
than the other operations.

A scan over the vector group operates in one invocation of the kernel,
whereas the full mux sub-group operates on more than one of these
vectorized invocations. As such, the individual invocations need access
to the previous invocations' scan results. It must therefore use mux
sub-group operations (a reduction and an exclusive scan) to propagate
the scan 'forward' across members of the sub-group.