[compiler] Vectorize to any required sub-group size
This updates users of the vectorizer to set the desired vectorization
factor according to the kernel's required sub-group size, if present. It
does so using a common method, to ensure consistent behaviour.This
helper method assumes that the current underlying 'mux' sub-group size
is 1, as we haven't got any in-tree targets that make use of anything
larger than that.

All targets consider the required sub-group size less than they do
`optnone` or `-cl-wfv=never` options. This can be revisited in the
future if users desire different behaviour.