[compiler] Preserve 'entry point' on vectorized kernels
This fixes a long-standing hack in the vectorizer where it would strip
the vectorized kernels of their 'kernel' status.

This wasn't anything to do with vectorization but was a legacy behaviour
owing to the fact that the vectorizer used to replace the old functions.
The `WorkItemLoopsPass` would in turn wrap every kernel it came across,
and we didn't want it wrapping up the old 'scalar' kernel since we'd
never call it: this would negatively impact compile time for no reason.

However, with the revised sub-groups model and to handle the case where
the target is not using degenerate sub-groups, we *do* want to wrap up
the unvectorized scalar kernel as it is useful as a fallback in the same
way the degenerate sub-group kernel is.

To avoid compiling unnecessary kernel wrappers, the WorkItemLoopsPass
has been taught to avoid wrapping scalar kernels if they don't use
sub-group builtins, or if they're degenerate, or if the kernel has a
required sub-group size. This should preserve the old behaviour in the
vast majority of cases.

The `VerifyReqdSubGroupSizeSatisfiedPass` must now be run later in the
pipeline, after the `WorkItemLoopsPass` has run. This is because
immediately after vectorization the old 'scalar' kernel appears as
though the required sub-group size has not been met. It is only once the
`WorkItemLoopsPass` has run and vectorized and unvectorized kernels have
been merged that we have our 'final' kernel forms, and we can glean
whether the compiler correctly satisfied the constraints.

We could alternatively make the pass look at all the vectorized forms of
each kernel and verify that *at least one* form of each kernel has met
the requirement, but this ties its behaviour more closely with the
vectorizer's output, whereas currently it's fairly agnostic: it runs
over all kernel entry points.