[riscv] Vectorize sub-group functions to a device size
This changes how the riscv/refsi targets handle sub-group sizes. The
riscv/refsi targets now automatically vectorize kernels/functions
containing explicit use of sub-groups to one of the device's advertised
sub-group widths.

These targets no longer run the `DegenerateSubGroupPass`!

The rationale for this is as follows: The SYCL 2020 spec has
`info::device::sub_group_sizes` which returns "the" list of sub-group
sizes reported by the device. The CTS expects that if it runs a kernel
which (for example) returns the kernel's sub-group size to the host,
then that size is found in that same device list.

This rules out the use of degenerate sub-groups. We'd have to list all
the sub-group (read: work-group) sizes from 1 to 1024, which is useless
for users. All of those sub-group sizes could then, in turn, be required
by a user, and the compiler would effectively be unable to satisfy. A
user requiring a sub-group size of 244 would lead to a very strange
situation where the compiler must create a degenerate sub-groups kernel
and would then just hope that the user uses a work-group size of 244.
Anything else would be invalid. This is another bad user experience.

The vectorizer can be asked to return a suitable vectorization factor
for a kernel, given its use of sub-groups and the device's list of
sub-group sizes. It is quite rudimentary in its heuristic and doesn't
choose between legal powers of two: the first one is taken as the
preferred list.

I've filed a ticket for clarification/questions about sub-groups and
degenerate sub-groups against the SYCL spec, but I think this is a
suitable "workaround" for now, or even for the long term. The behaviour
is arguably more intuitive for users.