[compiler] Add work-group scan support in vecz/work-item-loops
This completes the set of work-group collective functions we can support
via the 'alternative' pathway through the vectorizer and the
work-item-loops pass. They join broadcasts and reductions, which were
added a while back.

The scans are implemented using a linear work-item loop, accumulating
with each iteration of the 'X' loop. The vectorizer vectorizes them like
the sub-group scans, so the accumulator used is just the
current/previous value of scan operation so far. The vectorized
work-group merges this with the partial scan over the vector group.

The 'scans-only' parameter to the replace-wgc pass has been removed. The
pass is either run or it isn't run, depending on what the target wants
to do to support the work-group collective operations. Only the RefSi G1
WIPT example uses the old pass, though it should still be tested in LIT.

In unofficial local testing, these scans appear to be around 10% faster
to complete the associated SYCL-CTS tests. This requires more
investigation, however.