[compiler] Ensure createLoop always creates a loop
This simplifies the `createLoop` API so that users can rely on it having
generated a loop CFG structure. It correspondingly removes the
`allowUnroll` option.

Now, even when the loop bounds are constant and wouldn't generate a loop
(i.e., 0 or 1 iterations), `createLoop` will generate a loop. Users can
rely on standard LLVM passes to eliminate/simplify the loop.