[vecz] Packetize sub-group shuffles with uniform indices
This extends fixed-width vectorization capabilities to
`__mux_sub_group_shuffle` builtins, but only those with uniform indices
(where the shuffle index is the same for all invocations in the
sub-group). This accounts for the majority of those tested by the
SYCL-CTS. Support for varying indices will come down the line, once the
other shuffles are covered under similar conditions.

The existing sub-group LIT tests have been split by operation, as they
are expected to grow significantly to cover all of the different
conditions we can vectorize under.