[vecz] Fix vectorization of sub-group broadcasts
Forgetting to normalize the broadcast ID back into the 'mux' range would
produce incorrect results on an implementation where the mux sub-groups
were non-trivial, or where bounds were checked. This is, as such, not
currently a visible problem in the default implementation of these
builtins, but rather a theoretical one.