[multi_llvm] Move createBinOpForRecurKind helper
This hasn't belonged in multi_llvm for a long time. It's now just a
reusable compiler helper function.

This commit moves it into pass_functions.h, which itself is a bit
problematic as it's such a generic concept for a file. But it's better
than living in multi_llvm and we'll need to sort out 'pass_functions.h'
at some point in the future anyway.