[vecz] Fix SquashSmallVectorsPass on larger vectors
This commit fixes the optimization when it is faced with
vectors/integers (of the same size) whose bit-width is greater than 64
bits in size, or whose vector elements are at least 32 bits in size.

When the elements were at least 32 bits in size, the logic would exhibit
UB and shift a 32-bit integer (`1`) left by 32. This was leading to
incorrect code generation (the masks would come out as `0` rather than
the intended `0xFFFF...`).

Furthermore, when the vectors are greater than 64 bits in size, the
masking logic would break down as LLVM can only be given constants who
are representable by `uint64_t` and so we wouldn't be able to correctly
represent the mask.

At the same time, the the LIT test has been refactored to test only the
pass, as the output of the packetizer is largely uninteresting for this
test case.