[vecz] Provide more context when CFG conversion fails
This should help debugging (using `--pass-remarks-missed=vecz`) why
control-flow conversion failed to apply masks to the CFG.  The previous
diagnostics would only print the name of the function that couldn't be
converted, but not any more specific information.

This commit adds an extra level of information via a 'note', which is
optionally printed on the line after the main diagnostic.

This is not intended to be a full solution to better vecz diagnostics,
but a good first step.