[compiler] Handle scalable structs as barrier live variables
The work-item loops pass would crash when faced with a live variable
whose type was a struct containing scalable vectors.

We aren't legally allowed to store a struct type containing a mixture of
scalable and fixed types (`{ <vscale x 1 x i8>, i8 }`) so we decompose
such types into their constituent elements and store each individually.
Note that scalable elements are stored in the scalable part of the live
variables struct, and fixed elements are stored in the fixed part; in
that way they are treated as if they were never a struct to begin with.

Note that there may be a future optimization possible here where we
store all-scalable structs "whole", but this isn't currently a priority.

Note that this problem doesn't currently surface in the default pipeline
in the main branch, because we only end up with scalable vectors when we
vectorize as such, and we don't currently scalably vectorize any IR
that's known to contain struct types, at least not in a way that creates
a struct containing scalable vectors; see the new negative scalable vecz
test as an example.

The plan is to start allowing this when we improve the vectorization of
`cmpxhg` instructions. This should also improve the codegen for these
structures; see the new fixed-length vecz test for an example of the
poor codegen currently emitted.