[compiler] Remove the RemoveFencesPass
This pass was originally introduced back in LLVM 7 to account for the
fact that some of our targets couldn't handle the newly added `fence`
instruction. Removing them bypassed the issue.

While it is not semantically correct to remove fence instructions from
LLVM IR in general, we were relying on the fact that the memory model is
under-specified in OpenCL 1.2 and so could assume that fences were
redundant.

The ENORMOUS WARNING above this scheduling of this pass should be reason
enough not to run it. It has not been necessary for several LLVM
versions - it is claimed it is only in the host pipeline to ensure it's
tested. This is not a good enough reason to keep a dodgy unnecessary
pass.

There aren't enough good reasons to keep it around unused in the
codebase (the pass is trivial to implement by a downstream target if
they did require it) so this commit removes the pass.