[NFC] Change startswith/endswith to starts_with/ends_with.
For consistency with C++20, LLVM renamed startswith to starts_with, and
likewise for endswith. startswith and endswith are still available as
wrappers that do nothing but call starts_with and ends_with, but as of
LLVM 18 have been deprecated. As the oldest LLVM we support (LLVM 16)
already had starts_with and ends_with, we can just use that
unconditionally.