[vecz] Enable XFAIL tests
These two tests were actually passing in terms of the vectorizer, but
were then using a FileCheck invocation which failed, because there were
no CHECKs.

The two tests were originally testing that we couldn't packetize or
scalarize really wide vectors. However, we can. So this change updates
the tests to test the positive cases instead.