[vecz] Ensure inactive lanes don't contribute to branch conditions
In divergent blocks, we must ensure that masked-out - or inactive -
work-items don't contribute a 'true' value towards the branch condition.
Masked-out values may be poison or undef, such as those coming from a
masked-load operation.

Note that this can happen even with ostensibly 'uniform' values, as our
uniform/divergent analysis isn't capable of discerning an
unconditionally uniform value from one that's uniform but produces
poison/garbage when masked out.

To work around this, we ensure that branch conditions in divergent
blocks are masked with the block's entry mask. This should ensure that
only lanes that are meant to be active at that point contribute their
values.

This is likely not a 100% complete fix, as we don't perform this on
uniform blocks or loops. There are tests with 'uniform' loops which are
unconditionally entered despite no work-items being active. As such, if
we mask the loop exit condition with the entry mask, it will never be
true, and loop will never exit. This should be good enough to fix some
known regressions - the more correct fix would likely involve a lot more
work.

Note also that it is pessimistic in many of the vecz test changes. Some
(unmasked) uniform conditions don't need this applied, as they truly
always produce the same value even on inactive lanes. We don't have to
tools to distinguish these values from the ones that do need masking.