[NFC] Pick a style for const.
In cases where const may appear before or after the type, LLVM style is
to place it before the type, and we have many more instances ourselves
where we place it before the type than we do where we place it after.
This commit sets that style in .clang-format and does a reformat across
the code base.