[NFC] Add const.
Variables that can be defined as const are defined as const. This also
revealed that some loop iteration variables made pointless copies; these
have been changed to use references instead. Some variables also no
longer need to be captured in lambdas as a result of this change, so
they have been removed from the capture lists.