[compiler] Improve analysis of 'true uniform' values
This commit extends the Uniform Value Analysis with a fourth kind of
uniformity: "true" uniformity. This represents a value that is uniform
on both active and inactive lanes. The old class of "uniform" has been
renamed to "active" uniformity to clarify this.

The analysis pass has been extended with a method to query whether a
value is truly uniform. It processes this recursively on demand and
caches the result. This is because the initial analysis run works from
varying "roots" and marks all dependent values as varying - uniform
values aren't handled at all. Rather than negatively affecting the
performance of all Uniform Value Analysis runs, this on-demand method
keeps costs the same except for users that need to query uniformity.

The query is still conservative, but less so. This can be seen in the
test changes, where some cases which were previously conservatively
handled as possibly varying/active uniform are now seen as truly
uniform.