Add CA_ENABLE_TOOLS to dictate whether tools are built or not
We are not using oneapi-construction as a way of including externally
but this can lead to target noise. This change allows the tools not to
be added as a target - although some will be enabled if CA_ENABLE_TESTS
is enabled.

Also deleted modules/mux/targets/host/tools as this was not included in
any CMakelists.txt.