Upgrade to clang-tidy-17.
* misc-include-cleaner is disabled globally because at the moment, it
  requires annotations in external headers we have no control over.
* misc-use-anonymous-namespace is disabled globally because it goes
  against LLVM style, and the warning is triggered by use of LLVM macros
  that we have no control over.
* modernize-macro-to-enum is disabled globally because it also warns on
  macros that should be usable in preprocessor conditions.

Everything else either has the code adjusted according to the intent of
the warning, or has the warning suppressed as appropriate. Warnings are
suppressed for false positives (intentionally unused function return
values, and two instances of clang-tidy warning about undefined code
that is actually well-defined), and for code that we cannot improve
without a bigger refactor (unused function return values in places where
we have no way to indicate an error).