[vecz] Fix dropped value analysis in OffsetInfo
There were several cases in the OffsetInfo analysis where the OffsetInfo
was told to copy the stride information from another value (typically
one of the instruction's source operands) but the bitmask was not
updated. This meant that the value analysis - which works backwards from
the leaf instruction - would "stop" at that point, and become incorrect
as it would not take into account the range of values that the source
may have.

This was problematic because the analysis would come to the conclusion
that a value has a safe range of values and a valid stride, when in fact
it was truncated from a larger value before being zero-extended, and
thus divergent.

This was found through a segfault/invalid access in one of ArrayFire's
CannyEdgeDetector kernels.

Oftentimes we want to copy the bitmask *and* stride information over, so
a convenience function has been introduced. Other times we need to be
careful to update the bitmask correctly before copying the stride
information.