Keep debug information.
DeleteDebugInfoInstructions has a comment explaining that it intends to
remove invalid debug info instructions, but the ones that it is removing
seem perfectly valid. Given that Fraser recently removed other
generation of invalid debug info instructions, hopefully this means we
no longer have invalid debug info that we need to clean up.