[vecz] Clear getelementptr flags.
If a getelementptr has inbounds/nusw/nuw flags, it is valid for the
result to be out of bounds or to wrap so long as the result is not used
to access any object. However, if we vectorize, multiple elements will
all be loaded based on the first element's address, and in that we may
add accesses, so we need to clear the inbounds/nusw/nuw flags unless we
can prove they are still valid.