Removal of Vulkan API support
This removes Vulkan API support from the construction kit, including
all testing and documentation - see
https://github.com/uxlfoundation/oneapi-construction-kit/blob/rfcs/rfc-0003.md
for more information.

Note that some code which is only being used by Vulkan remains after this
commit. This will be reviewed in a later commit.

Note that the next release should increase the version to 5.0.0 after
this breaking change.