LLVM 20: Update for getFirstNonPHIOrDbg.
LLVM 20 updates getFirstNonPHIOrDbg() to return an iterator rather than
an Instruction *. We can use &* to reliably get an Instruction * across
LLVM versions.

The variable name insert_point suggests that we should instead change
the type of the variable to be an iterator, but despite the name, it is
not used merely as an insert point. This should be revisited in the
future but may result in changes in behavior that should probably be
kept separate from any compatibility fixes.