Increase CMake minimum version.
Despite declaring a minimum CMake version of 3.16 at the top level, we
were using CMake 3.19+ features without checking for support for them
already. As no 3.19 version of CMake is available at PyPI for testing,
bump the minimum to 3.20 instead which we can check to work.

Propagate this change throughout the project to fix compatibility with
CMake 4.0, which will raise errors on too old cmake_minimum_required
versions.