[NFC] Address LLVM 21's BinOp additions in a better way.
The earlier change to restore compatibility with LLVM 21 required
suppressing compiler warnings and made it so that we would not be
alerted when we need to account for future BinOp additions.

This change moves the complexity of dealing with the different LLVM
versions from vectorization_context.cpp into multi_llvm, and handles it
in a way that does not require suppressing compiler warnings.