Remove outdated test.
undef_debug_info.ll was originally added to test that we did not
generate debug info with undef, because we were unable to handle that
properly. Later on, when we were able to handle that properly, I removed
the logic for keeping undef out of debug info in #375, at which point
this test no longer tested anything useful. This PR finally removes it.